Task development

+++ Inc/temp_sensor.h
@@ -10,6 +10,6 @@
 uint8_t DS18B20_Start(void);
 void DS18B20_Write(uint8_t data);
 uint8_t DS18B20_Read(void);
-float DS18B20_GetCelsiusTemp(void);
+int DS18B20_GetCelsiusTemp(void);
 
 #endif
